// frontend/src/app/generated/api/model/profileDto.ts
/**
 * Yaam.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CertificationDto } from './certificationDto';
import { LanguageDto } from './languageDto';
import { ProfileLinkDto } from './profileLinkDto';
import { WorkExperienceDto } from './workExperienceDto';
import { CustomFieldDto } from './customFieldDto';
import { EducationDto } from './educationDto';


export interface ProfileDto { 
    id: string;
    firstName: string | null;
    lastName: string | null;
    email: string | null;
    phone: string | null;
    location: string | null;
    summary: string | null;
    workExperiences: Array<WorkExperienceDto>;
    educations: Array<EducationDto>;
    skills: Array<string>;
    languages: Array<LanguageDto>;
    certifications: Array<CertificationDto>;
    links: Array<ProfileLinkDto>;
    customFields: Array<CustomFieldDto>;
}